refactor(dogfood): remove not required items
