+++ x/dogfood/client/cli/tx.go
@@ -2,24 +2,12 @@ package cli
 
 import (
 	"fmt"
-	"time"
 
 	"github.com/spf13/cobra"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	// "github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/ExocoreNetwork/exocore/x/dogfood/types"
-)
-
-var (
-	DefaultRelativePacketTimeoutTimestamp = uint64(
-		(time.Duration(10) * time.Minute).Nanoseconds(),
-	)
-)
-
-const (
-	flagPacketTimeoutTimestamp = "packet-timeout-timestamp"
-	listSeparator              = ","
 )
 
 // GetTxCmd returns the transaction commands for this module
