+++ x/dogfood/types/keys.go
@@ -6,11 +6,4 @@ const (
 
 	// StoreKey defines the primary module store key
 	StoreKey = ModuleName
-
-	// RouterKey defines the module's message routing key
-	RouterKey = ModuleName
 )
-
-func KeyPrefix(p string) []byte {
-	return []byte(p)
-}
